Enable to run mencoder(static-linked and single thread).
The reason that unitrans cannot run mencoder is that mencoder use the
'ffreep' instruction in the InitTimer function. But 'ffreep' instrucion's
emit function is NULL. So, the unitrans call its translation function will
call the 0 address. So the DBT system crashed.

By specifing ffreep's emit funtion as emit_normal, solve the bug.